requirements-dev.txt
-r requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.1
pytest==3.0.6
